practice/03_agent/pyproject.toml
[project]
name = "03-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.11",
]
